Refactor project layout for pyproject.toml and setup.cfg
