+++ .gitignore
@@ -9,3 +9,6 @@
 __pycache__
 .mypy_cache
 .pytest_cache
+
+# Builds
+build
