+++ example/config_sample
@@ -1,0 +1,58 @@
+#<<<<< SSH Config file managed by sshclick >>>>>
+
+#-------------------------------------------------------------------------------
+#@group: network
+#@desc: Network devices in my lab
+#@info: user='admin' password='password'
+#@info: Not really, but for demo its ok :)
+#-------------------------------------------------------------------------------
+Host net-switch1
+    hostname 10.1.1.1
+
+Host net-switch2
+    hostname 10.1.1.2
+
+Host net-switch3
+    hostname 10.1.1.3
+
+#@host: Same user on all hosts matching this pattern
+Host net-*
+    user admin
+
+
+#-------------------------------------------------------------------------------
+#@group: jumphost
+#@desc: Edge-server / SSH bastion
+#@info: Used for jump-proxy from intnet to internal lab servers
+#-------------------------------------------------------------------------------
+Host jumper1
+    hostname 123.123.123.123
+    user master
+    port 1234
+
+
+#-------------------------------------------------------------------------------
+#@group: lab-servers
+#@desc: Testing/Support servers
+#@info: Some [red]important[/] detail here!
+#@info: We can have color markups in descriptions and info lines
+#-------------------------------------------------------------------------------
+#@host: adding info line
+Host lab-serv1
+    hostname 10.10.0.1
+    user admin
+
+Host lab-serv2
+    hostname 10.16.141
+
+Host server-behind-lab
+    hostname 10.30.0.1
+    user testuser
+    port 1234
+    proxyjump lab-serv1
+    localforward 8080 127.0.0.1:80
+
+Host lab-*
+    user user123
+    proxyjump jumper1
+
