+++ setup.cfg
@@ -1,0 +1,38 @@
+[metadata]
+name = sshclick
+version = attr: sshclick.VERSION
+author = Karlo Tisaj
+author_email = karlot@gmail.com
+description = SSH Config manager
+license = MIT
+url = https://github.com/karlot/sshclick
+classifiers =
+    Environment :: Console
+    Operating System :: POSIX :: Linux
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
+    
+
+[options]
+zip_safe = False
+include_package_data = True
+python_requires = >=3.7, <4
+install_requires = 
+    click>=8.1.3
+    prettytable>=3.2.0
+    rich>=12.5.1
+
+[options.entry_points]
+console_scripts =
+    sshc = sshclick.main:cli
+
+[options.extras_require]
+test = 
+    pytest
+
+[options.package_data]
+sample = 
+    example/config_sample
