+++ sshclick/__init__.py
@@ -1,0 +1,1 @@
+VERSION = "0.4.2"
