+++ sshclick/sshc/ssh_group.py
